Create block based lookahead.
Agents now look ahead and try to grab a block that will be needed in the
future. Does not handle blocks being grabbed at the same time. Currently
they just drop the blocks again if it is too many for the current subset
of blocks looked at. Eventually one will be the only one and be taking
it to the dropzone.